refactoring(ci): do not use uint #46

+++ app/apps/project/adapter/status_adapter.go
@@ -8,10 +8,10 @@ import (
 
 // TaskStatusDto - task status dto.
 type TaskStatusDto struct {
-	ID        uint                     `json:"id"`
+	ID        int64                    `json:"id"`
 	Title     string                   `json:"title"`
-	ProjectId uint                     `json:"project_id"`
-	Index     uint                     `json:"index"`
+	ProjectId int64                    `json:"project_id"`
+	Index     int64                    `json:"index"`
 	Config    *domain.TaskStatusConfig `json:"config"`
 }
