+++ app/apps/user/adapter/token_dto.go
@@ -10,7 +10,7 @@ import (
 // It is used to expose token information to the client (e.g., via an API)
 // while hiding sensitive fields like the actual refresh token string.
 type UserTokenDto struct {
-	Id               uint       `json:"id"`                 // Unique ID of the token record
+	Id               int64      `json:"id"`                 // Unique ID of the token record
 	Device           string     `json:"device"`             // Device or client info associated with the token
 	RefreshExpiresAt time.Time  `json:"refresh_expires_at"` // Expiration time of the refresh token
 	Revoked          bool       `json:"revoked"`            // Flag indicating if the token was revoked
